Use a database!!!
This is a huge change but it simplifies things immensely.
So much of this code is just managing what data goes where, and keeping it normalized. When I put it into a normalized database, everything is so much cleaner!

A lot of cleanup is necessary, but a test passes, so I'm committing it.